+++ test/test_listing.py
@@ -36,7 +36,7 @@ class AddBaseEntryTestCase(unittest.TestCase):
         self.listing = Listing()
         self.item_name = "Aldi"
         self.item_value = 66.6
-        self.item_date = "11-08"
+        self.item_date = "2000-11-08"
         self.item_category = "Groceries"
         self.listing.add_entry(
             BaseEntry(self.item_name, self.item_value, self.item_date),
@@ -46,9 +46,9 @@ class AddBaseEntryTestCase(unittest.TestCase):
         self.assertEqual(
             self.listing.prettify(), '\n'.join([
                 "{1:^{0}}".format(CategoryEntry.TOTAL_LENGTH, "Listing"),
-                "Name               Value    Date  ID ",
-                "Groceries             66.60" + 10 * " ",
-                "  Aldi                66.60 11-08   0"
+                "Name               Value    Date     ID ",
+                "Groceries             66.60" + 13 * " ",
+                "  Aldi                66.60 00-11-08   0"
             ]))
 
     def test_str_no_eid(self):
@@ -57,10 +57,10 @@ class AddBaseEntryTestCase(unittest.TestCase):
             self.listing.prettify(),
             '\n'.join([
                 "{1:^{0}}".format(CategoryEntry.TOTAL_LENGTH, "Listing"),
-                "Name               Value    Date ",
+                "Name               Value    Date    ",
                 # TODO: fix this; category entry line has to be shorter
-                "Groceries             66.60          ",
-                "  Aldi                66.60 11-08"
+                "Groceries             66.60             ",
+                "  Aldi                66.60 00-11-08"
             ]))
         BaseEntry.SHOW_EID = True
 
@@ -72,28 +72,28 @@ class SortCategoryEntriesTestCase(unittest.TestCase):
     def setUp(self):
         self.listing = Listing()
         for c, v in zip("ab", [20, 10]):
-            self.listing.add_entry(BaseEntry("foo", v, "01-01"), c)
+            self.listing.add_entry(BaseEntry("foo", v, "2000-01-01"), c)
 
     def test_sort_by_name(self):
         self.assertEqual(
             self.listing.prettify(category_sort="name"), '\n'.join([
                 "{1:^{0}}".format(CategoryEntry.TOTAL_LENGTH, "Listing"),
-                "Name               Value    Date  ID ",
-                "A                     20.00" + 10 * " ",
-                "  Foo                 20.00 01-01   0",
-                "B                     10.00" + 10 * " ",
-                "  Foo                 10.00 01-01   0",
+                "Name               Value    Date     ID ",
+                "A                     20.00" + 13 * " ",
+                "  Foo                 20.00 00-01-01   0",
+                "B                     10.00" + 13 * " ",
+                "  Foo                 10.00 00-01-01   0",
             ]))
 
     def test_sort_by_value(self):
         self.assertEqual(
             self.listing.prettify(), '\n'.join([
                 "{1:^{0}}".format(CategoryEntry.TOTAL_LENGTH, "Listing"),
-                "Name               Value    Date  ID ",
-                "B                     10.00" + 10 * " ",
-                "  Foo                 10.00 01-01   0",
-                "A                     20.00" + 10 * " ",
-                "  Foo                 20.00 01-01   0",
+                "Name               Value    Date     ID ",
+                "B                     10.00" + 13 * " ",
+                "  Foo                 10.00 00-01-01   0",
+                "A                     20.00" + 13 * " ",
+                "  Foo                 20.00 00-01-01   0",
             ]))
 
 
@@ -102,7 +102,7 @@ class AddNegativeBaseEntryTestCase(unittest.TestCase):
         self.listing = Listing()
         self.item_name = "Aldi"
         self.item_value = -66.6
-        self.item_date = "11-08"
+        self.item_date = "2000-11-08"
         self.item_category = "Groceries"
         self.listing.add_entry(
             BaseEntry(self.item_name, self.item_value, self.item_date),
@@ -112,9 +112,9 @@ class AddNegativeBaseEntryTestCase(unittest.TestCase):
         self.assertEqual(
             self.listing.prettify(), '\n'.join([
                 "{1:^{0}}".format(CategoryEntry.TOTAL_LENGTH, "Listing"),
-                "Name               Value    Date  ID ",
-                "Groceries             66.60" + 10 * " ",
-                "  Aldi                66.60 11-08   0"
+                "Name               Value    Date     ID ",
+                "Groceries             66.60" + 13 * " ",
+                "  Aldi                66.60 00-11-08   0"
             ]))
 
 
@@ -123,7 +123,7 @@ class AddBaseEntryWithoutCategoryTestCase(unittest.TestCase):
         self.listing = Listing()
         self.item_name = "Aldi"
         self.item_value = 66.6
-        self.item_date = "11-08"
+        self.item_date = "2009-11-08"
         self.listing.add_entry(
             BaseEntry(self.item_name, self.item_value, self.item_date),
             category_name=CategoryEntry.DEFAULT_NAME)
@@ -139,7 +139,7 @@ class AddTwoBaseEntriesTestCase(unittest.TestCase):
         self.item_a_value = 66.6
         self.item_b_value = 10.01
         self.item_category = "Groceries"
-        self.date = "11-11"
+        self.date = "2008-11-11"
         self.listing.add_entry(
             BaseEntry("Aldi", self.item_a_value, self.date), self.item_category)
         self.listing.add_entry(
@@ -156,13 +156,13 @@ class ListingFromElementsTestCase(unittest.TestCase):
     def setUp(self):
         self.name = "Dinner for one"
         self.value = 99.9
-        self.date = "12-31"
+        self.date = "1980-12-31"
         self.listing = Listing.from_elements(
             [dict(name=self.name, value=self.value, date=self.date, eid=0)],
             default_category=CategoryEntry.DEFAULT_NAME)
 
     def test_contains_an_entry(self):
-        self.assertIn(self.date, self.listing.prettify())
+        self.assertIn(self.date[2:], self.listing.prettify())
 
     def test_category_item_names(self):
         parsed_listing_entry_names = list(self.listing.category_entry_names)
@@ -181,20 +181,20 @@ class PrettifyListingsTestCase(unittest.TestCase):
                 1: {
                     "name": "food",
                     "value": -100.01,
-                    "date": "03-03",
+                    "date": "2000-03-03",
                     "category": "groceries"
                 },
                 999: {
                     "name": "money",
                     "value": 299.99,
-                    "date": "03-03"
+                    "date": "2000-03-03"
                 }
             },
             "recurrent": {
                 42: [{
                     "name": "gold",
                     "value": 4321,
-                    "date": "01-01",
+                    "date": "2000-01-01",
                     "category": "bank"
                 }]
             }
@@ -204,15 +204,15 @@ class PrettifyListingsTestCase(unittest.TestCase):
         self.assertEqual(
             prettify(
                 elements_copy, default_category=CategoryEntry.DEFAULT_NAME),
-            "              Earnings                |               Expenses               \n"  # noqa
-            "Name               Value    Date  ID  | Name               Value    Date  ID \n"  # noqa
-            "Unspecified          299.99           | Groceries            100.01          \n"  # noqa
-            "  Money              299.99 03-03 999 |   Food               100.01 03-03   1\n"  # noqa
-            "Bank                4321.00           | \n"  # noqa
-            "  Gold              4321.00 01-01  42 | \n"  # noqa
-            "=============================================================================\n"  # noqa
-            "Total               4620.99           | Total                100.01          \n"  # noqa
-            "Difference          4520.98          ")
+            "                Earnings                 |                 Expenses                \n"  # noqa
+            "Name               Value    Date     ID  | Name               Value    Date     ID \n"  # noqa
+            "Unspecified          299.99              | Groceries            100.01             \n"  # noqa
+            "  Money              299.99 00-03-03 999 |   Food               100.01 00-03-03   1\n"  # noqa
+            "Bank                4321.00              | \n"  # noqa
+            "  Gold              4321.00 00-01-01  42 | \n"  # noqa
+            "===================================================================================\n"  # noqa
+            "Total               4620.99              | Total                100.01             \n"  # noqa
+            "Difference          4520.98             ")
         # Assert that original data was not modified
         self.assertDictEqual(elements, elements_copy)
 
@@ -221,13 +221,13 @@ class PrettifyListingsTestCase(unittest.TestCase):
                 elements_copy,
                 default_category=CategoryEntry.DEFAULT_NAME,
                 category_percentage=True),
-            "              Earnings                |               Expenses               \n"  # noqa
-            "Name               Value        %     | Name               Value        %    \n"  # noqa
-            "Unspecified          299.99   6.5     | Groceries            100.01 100.0    \n"  # noqa
-            "Bank                4321.00  93.5     | \n"  # noqa
-            "=============================================================================\n"  # noqa
-            "Total               4620.99           | Total                100.01          \n"  # noqa
-            "Difference          4520.98          ")
+            "                Earnings                 |                 Expenses                \n"  # noqa
+            "Name               Value           %     | Name               Value           %    \n"  # noqa
+            "Unspecified          299.99      6.5     | Groceries            100.01    100.0    \n"  # noqa
+            "Bank                4321.00     93.5     | \n"  # noqa
+            "===================================================================================\n"  # noqa
+            "Total               4620.99              | Total                100.01             \n"  # noqa
+            "Difference          4520.98             ")
 
     def test_prettify_stacked_layout(self):
         elements = {
@@ -235,35 +235,36 @@ class PrettifyListingsTestCase(unittest.TestCase):
                 2: {
                     "name": "shirt",
                     "value": -199,
-                    "date": "04-01",
+                    "date": "2000-04-01",
                     "category": "clothes",
                 },
                 3: {
                     "name": "lunch",
                     "value": -20,
-                    "date": "04-01",
+                    "date": "2000-04-01",
                     "category": "food",
                 }
             },
             "recurrent": {}
         }
+        self.maxDiff = None
         self.assertEqual(
             prettify(elements, stacked_layout=True), "\
-              Earnings               " + "\n\
-Name               Value    Date  ID " + "\n\
-=====================================" + "\n\
-Total                  0.00          " + """
+                Earnings                " + "\n\
+Name               Value    Date     ID " + "\n\
+========================================" + "\n\
+Total                  0.00             " + """
 
 """ + "\
-              Expenses               " + "\n\
-Name               Value    Date  ID " + "\n\
-Food                  20.00          " + "\n\
-  Lunch               20.00 04-01   3" + "\n\
-Clothes              199.00          " + "\n\
-  Shirt              199.00 04-01   2" + "\n\
-=====================================" + "\n\
-Total                219.00          " + "\n\
-Difference          -219.00          ")
+                Expenses                " + "\n\
+Name               Value    Date     ID " + "\n\
+Food                  20.00             " + "\n\
+  Lunch               20.00 00-04-01   3" + "\n\
+Clothes              199.00             " + "\n\
+  Shirt              199.00 00-04-01   2" + "\n\
+========================================" + "\n\
+Total                219.00             " + "\n\
+Difference          -219.00             ")
 
 
 if __name__ == '__main__':