+++ test/utils.py
@@ -14,4 +14,4 @@ class Client(clients.LocalServerClient):
             configuration=config.Configuration(),
             sinks=clients.Client.Sinks(f, f))
 
-        self.proxy._period_kwargs["data_dir"] = None
+        self.proxy._pocket_kwargs["data_dir"] = None